requirements.txt
streamlit
google-genai
psycopg2-binary
pgvector
python-dotenv
PyPDF2
python-docx
pandas
streamlit-option-menu
plotly
streamlit-authenticator
requests
